moved test app to seperate package
