+++ testutil/utils/accounts.go
@@ -3,6 +3,7 @@ package utils
 import (
 	"cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	app "github.com/dymensionxyz/dymension-rdk/testutil/app"
 	"github.com/tendermint/tendermint/crypto/ed25519"
 
 	ibctransfertypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
@@ -30,7 +31,7 @@ func createRandomAccounts(accNum int) []sdk.AccAddress {
 
 // AddTestAddrs constructs and returns accNum amount of accounts with an
 // initial balance of accAmt in random order
-func AddTestAddrs(app *App, ctx sdk.Context, accNum int, accAmt math.Int) []sdk.AccAddress {
+func AddTestAddrs(app *app.App, ctx sdk.Context, accNum int, accAmt math.Int) []sdk.AccAddress {
 	testAddrs := createRandomAccounts(accNum)
 	initCoins := sdk.NewCoins(sdk.NewCoin(app.StakingKeeper.BondDenom(ctx), accAmt))
 
@@ -41,7 +42,7 @@ func AddTestAddrs(app *App, ctx sdk.Context, accNum int, accAmt math.Int) []sdk.
 	return testAddrs
 }
 
-func InitAccountWithCoins(app *App, ctx sdk.Context, addr sdk.AccAddress, coins sdk.Coins) {
+func InitAccountWithCoins(app *app.App, ctx sdk.Context, addr sdk.AccAddress, coins sdk.Coins) {
 	err := app.BankKeeper.MintCoins(ctx, ibctransfertypes.ModuleName, coins)
 	if err != nil {
 		panic(err)
@@ -53,7 +54,7 @@ func InitAccountWithCoins(app *App, ctx sdk.Context, addr sdk.AccAddress, coins 
 	}
 }
 
-func FundModuleAccount(app *App, ctx sdk.Context, moduleName string, coins sdk.Coins) {
+func FundModuleAccount(app *app.App, ctx sdk.Context, moduleName string, coins sdk.Coins) {
 	if err := app.BankKeeper.MintCoins(ctx, ibctransfertypes.ModuleName, coins); err != nil {
 		panic(err)
 	}
